Image Analysis
This code helps to solve a common problem with photographers, which is plagiarism. Photographers have issues with individuals copying a subset of their original image just to use for their own objective. Google Image Search is a modern day solution for this kind of problem, but the code in this exercise gives you a general idea of how the program would work.